+++ noodles-vcf/src/header/record/value/map/info/number.rs
@@ -1,5 +1,3 @@
-use std::fmt;
-
 /// A VCF number describing the cardinality of a field.
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub enum Number {
@@ -32,18 +30,6 @@ impl Default for Number {
     }
 }
 
-impl fmt::Display for Number {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Count(n) => write!(f, "{n}"),
-            Self::AlternateBases => f.write_str("A"),
-            Self::ReferenceAlternateBases => f.write_str("R"),
-            Self::Samples => f.write_str("G"),
-            Self::Unknown => f.write_str("."),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -51,14 +37,5 @@ mod tests {
     #[test]
     fn test_default() {
         assert_eq!(Number::default(), Number::Count(1));
-    }
-
-    #[test]
-    fn test_fmt() {
-        assert_eq!(Number::Count(1).to_string(), "1");
-        assert_eq!(Number::AlternateBases.to_string(), "A");
-        assert_eq!(Number::ReferenceAlternateBases.to_string(), "R");
-        assert_eq!(Number::Samples.to_string(), "G");
-        assert_eq!(Number::Unknown.to_string(), ".");
     }
 }