+++ noodles-sam/CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Removed
 
-  * vcf/header/record/value/map/info/number: Remove parser.
+  * vcf/header/record/value/map/info/number: Remove parser (`fmt::FromStr`) and
+    formatter (`fmt::Display`).
+
+    This prevents coupling the domain model to a form of serialization.
 
 ## 0.63.0 - 2024-08-04